perf(graphql): skip peer hydration for id-only single relationships (#10062)
* perf(graphql): skip peer hydration for id-only single relationships

When a cardinality-one relationship query requests only the peer node's
`id` (no properties, metadata, or additional node fields), return the
peer ID already loaded on the parent instead of hydrating a full peer
Node via NodeDataLoader. This avoids redundant database work on
relationship-heavy queries such as cable-plan reports.

The resolver falls through to the existing NodeDataLoader / query_peers
paths whenever any other field, property, or metadata is requested, so
behavior is unchanged for all other queries.

* fix(graphql): skip id-only relationship shortcut for generic peers

The id-only single-relationship shortcut returns {"id": peer_id} without
the peer kind. For a relationship whose peer is a generic, the GraphQL
node field is an interface whose resolve_type needs the kind, so it fails
with "Unable to identify the type of the instance." Restrict the shortcut
to concrete peers; generic peers fall back to full hydration.


* perf(graphql): carry peer kind so generic id-only relationships skip hydration

The id-only single-relationship shortcut had to be disabled for generic
peers: it returned {"id": peer_id} without a kind, and a generic peer
resolves to a GraphQL interface whose resolve_type indexes the type
registry by kind. That excluded a large share of core-schema
relationships from the optimization.

The peer's concrete kind was available all along. The relationship query
already binds the peer vertex and returns peer.uuid, so returning
peer.kind alongside it costs no extra round trip. Thread that kind
through GroupedPeerNodes, PeerWithRelationshipMetadata and Relationship
so the preloaded stub on a node's GraphQL response can carry __kind__,
and let the shortcut serve generic peers from it.

The shortcut declines whenever the kind is absent and the resolver
hydrates as before, so correctness never depends on the enrichment.


* test(graphql): assert generic-peer id-only relationship skips hydration

Add a component test that exercises the id-only single-relationship shortcut
against a cardinality-one relationship whose peer is a generic. It monkeypatches
NodeDataLoader.load to raise, so a regression to the correct-but-slower
hydrating path fails loudly instead of passing silently.

* test(benchmark): add id-only single-relationship benchmark

Add test_query_rel_one_id_only mirroring test_query_rel_one but selecting
only the peer id (repository { node { id } }), so the id-only
single-relationship shortcut is exercised and validated by CodSpeed CI.

* update name, reduce comment

* consolidate tests

* reduce size of huge test file

---------
